+++ projects/ids-enterprise-ng/src/lib/datagrid/soho-datagrid.component.ts
@@ -1081,6 +1081,10 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
   @Output()
   expandrow = new EventEmitter<SohoDataGridToggleRowEvent>();
 
+  // This event is fired when a key is pressed
+  @Output()
+  keydown = new EventEmitter<SohoDataGridKeyDownEvent>();
+
   // This event is fired when edit mode is exited.
   @Output()
   exiteditmode = new EventEmitter<SohoDataGridEditModeEvent>();
@@ -1366,6 +1370,15 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
   clearDirty(): void {
     return this.ngZone.runOutsideAngular(() => {
       this.datagrid.clearDirty();
+    });
+  }
+
+  /**
+   * Commit the cell that's currently in edit mode.
+   */
+  commitCellEdit(): void {
+    return this.ngZone.runOutsideAngular(() => {
+      this.datagrid.commitCellEdit();
     });
   }
 
@@ -1776,6 +1789,16 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
     const event = { grid: this, ...args };
     this.ngZone.run(() => {
       this.expandrow.next(event);
+    });
+  }
+
+  /**
+   * Event fired after a key is pressed
+   */
+  private onKeyDown(e: JQuery.Event, args: SohoDataGridKeyDownArgs, response: Function) {
+    const event = { e, args, response };
+    this.ngZone.run(() => {
+      this.keydown.next(event);
     });
   }
 
@@ -2206,6 +2229,11 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
         this.onEditCell(editor);
       };
 
+      // Add the keydown callback.
+      this._gridOptions.onKeyDown = (e: JQuery.Event, args: SohoDataGridKeyDownArgs, response: Function) => {
+        this.onKeyDown(e, args, response);
+      };
+
       // Initialise any event handlers.
       this.jQueryElement
         .on('addrow', (e: any, args: SohoDataGridAddRowEvent) => { this.onRowAdd(args); })
@@ -2353,4 +2381,14 @@ export interface SohoDataGridToggleRowEvent extends SohoDataGridRowExpandEvent {
   // The data grid component originating the call.
   grid: SohoDataGridComponent;
   args?: any;
+}
+
+/**
+ * Details of the 'keydown' event
+ */
+export interface SohoDataGridKeyDownEvent {
+  // The data grid component originating the call.
+  e: JQuery.Event;
+  args?: SohoDataGridKeyDownArgs;
+  response?: Function;
 }