+++ projects/ids-enterprise-ng/src/lib/datagrid/soho-datagrid.d.ts
@@ -239,6 +239,11 @@ interface SohoDataGridOptions {
   onEditCell?: SohoDataGridEditCellFunction;
 
   /**
+  * Vetoable Key Down Callback
+  */
+  onKeyDown?: SohoDataGridKeyDownFunction;
+
+  /**
   * A callback function that fires when expanding rows.
   * To be used when expandableRow is true.
   * The function gets eventData about the row and grid and a response
@@ -341,6 +346,20 @@ interface SohoDataGridPostRenderCellArgs {
   api: SohoDataGridStatic;
 }
 
+/**
+ * The arguments object passed to the onKeyDown callback.
+ */
+interface SohoDataGridKeyDownArgs {
+  /** Info about the active cell. */
+  activeCell: any;
+
+  /** The row index. */
+  row: number;
+
+  /** The cell index. */
+  cell: number;
+}
+
 interface SohoDataGridEditCellFunctionArgs extends SohoDataGridPostRenderCellArgs {
   container: any;
   e: any;
@@ -392,6 +411,12 @@ type SohoDataGridPostRenderCellFunction = (
 
 type SohoDataGridEditCellFunction = (
   editor: any
+) => void;
+
+type SohoDataGridKeyDownFunction = (
+  e: JQuery.Event,
+  args: SohoDataGridKeyDownArgs,
+  response: Function
 ) => void;
 
 type SohoDataGridSourceFunction = (
@@ -1031,6 +1056,11 @@ interface SohoDataGridStatic {
   clearAllErrors(): void;
 
   /**
+   * Commit the cell that's currently in edit mode.
+   */
+  commitCellEdit(): void;
+
+  /**
    * Sets the status of a given row in the grid.
    *
    * @param idx - the row number (idx) of the row
@@ -1221,6 +1251,9 @@ interface SohoDataGridGroupable {
 
   // Type of aggregation.
   aggregator: SohoDataGridAggregator;
+
+  // Formatter for group row
+  groupRowFormatter?: SohoDataGridColumnFormatterFunction;
 }
 
 type SohoDataGridAggregator = 'sum' | 'max' | 'list' | string;
@@ -1302,23 +1335,29 @@ interface SohoDataGridEditModeEvent {
   row: number;
 
   /** Column number. */
-  cell: number,
+  cell: number;
 
   /** Row data */
-  item: any,
+  item: any;
 
   /** HTMLElement of the owning cell */
-  target: HTMLElement,
+  target: HTMLElement;
 
   /** The cell value. */
-  value: any,
+  value: any;
 
   /** The original cell value. */
-  oldValue: any,
+  oldValue: any;
 
   /** The column definition. */
-  column: SohoDataGridColumn,
+  column: SohoDataGridColumn;
 
   /** The cell editor object. */
-  editor: SohoDataGridCellEditor
+  editor: SohoDataGridCellEditor;
+}
+
+interface SohoDataGridKeyDownEvent {
+  e: JQuery.Event;
+  args: SohoDataGridKeyDownArgs;
+  response: Function;
 }