+++ projects/ids-enterprise-ng/src/lib/fileupload/soho-fileupload.component.ts
@@ -115,6 +115,10 @@ export class SohoFileUploadComponent implements AfterViewInit, OnDestroy {
   }
 
   ngOnDestroy() {
+    // call outside the angular zone so change detection isn't triggered by the soho component.
+    if (this.jQueryElement) {
+      this.jQueryElement.off();
+    }
     if (this.fileUpload) {
       this.fileUpload.destroy();
       this.fileUpload = null;