+++ projects/ids-enterprise-ng/src/lib/homepage/soho-homepage.module.ts
@@ -5,11 +5,13 @@ import { SohoWidgetComponent } from './soho-widget.component';
 import { SohoWidgetContentComponent } from './soho-widget-content.component';
 import { SohoWidgetHeaderComponent } from './soho-widget-header.component';
 import { SohoWidgetTitleComponent } from './soho-widget-title.component';
+import { SohoHomepageSizerDirective } from './soho-homepage-sizer.directive';
 
 @NgModule({
   imports:      [ CommonModule ],
   declarations: [
     SohoHomePageComponent,
+    SohoHomepageSizerDirective,
     SohoWidgetComponent,
     SohoWidgetContentComponent,
     SohoWidgetHeaderComponent,
@@ -17,6 +19,7 @@ import { SohoWidgetTitleComponent } from './soho-widget-title.component';
   ],
   exports:      [
     SohoHomePageComponent,
+    SohoHomepageSizerDirective,
     SohoWidgetComponent,
     SohoWidgetContentComponent,
     SohoWidgetHeaderComponent,